move systemOverlayStyle to Theme
