+++ sdk/python/packages/flet-core/src/flet_core/__init__.py
@@ -20,7 +20,7 @@ from flet_core.alert_dialog import AlertDialog
 from flet_core.alignment import Alignment
 from flet_core.animated_switcher import AnimatedSwitcher, AnimatedSwitcherTransition
 from flet_core.animation import Animation, AnimationCurve
-from flet_core.app_bar import AppBar, SystemOverlayStyle
+from flet_core.app_bar import AppBar
 from flet_core.audio import Audio
 from flet_core.audio_recorder import AudioEncoder, AudioRecorder
 from flet_core.badge import Badge
@@ -216,6 +216,7 @@ from flet_core.theme import (
     PageTransitionsTheme,
     PageTransitionTheme,
     ScrollbarTheme,
+    SystemOverlayStyle,
     TabsTheme,
     TextTheme,
     Theme,
